refine mcp test scripts
